+++ Tcp/src/main/resources/META-INF-DNS-SEEKER/META-INF/MANIFEST.MF
