+++ Tcp/src/main/resources/META-INF-DNS/META-INF/MANIFEST.MF
