--- tcp/server_linux/CMakeLists.txt
@@ -1,11 +1,0 @@
-cmake_minimum_required(VERSION 2.8)
-project(server_linux)
-
-find_library(PTHREAD_LIBRARY pthread)
-
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall -Wextra")
-
-set(SOURCE_FILES main.c)
-add_executable(server_linux ${SOURCE_FILES})
-
-target_link_libraries(server_linux ${PTHREAD_LIBRARY})
